feat: implement ownership control and role-based authorization

+++ Controllers/TodoController.cs
@@ -1,11 +1,13 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ToDoApi.DTOs;
+using System.Security.Claims;
+
 
 [Authorize]
 [ApiController]
 [Route("api/[controller]")]
-[Produces("application/json")] //informs swagger that API returns in json
+[Produces("application/json")] 
 public class TodoController : ControllerBase
 {
     private readonly ITodoService _service;
@@ -15,17 +17,25 @@ public class TodoController : ControllerBase
         _service = service;
     }
 
+
     /// <summary>
-    /// Retrieves all todo items.
+    /// Updates an existing todo item. (Admin and User)
     /// </summary>
-    /// <response code="200">Returns the list of todo items.</response>
-    /// <response code="401">Unauthorized if the token is missing or invalid.</response>
+    /// <response code="204">If the update was successful.</response>
+    /// <response code="401">If the user is not authenticated.</response>
+    /// <response code="403">If the user does not have permission to update this item.</response>
+    /// <response code="404">If the item to update does not exist.</response>
     [HttpGet]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> GetAll()
     {
         var todos = await _service.GetAllAsync();
+
+        if (CurrentUserRole != "Admin")
+        {
+            todos = todos.Where(x => x.UserId == CurrentUserId).ToList();
+        }
 
         var response = todos.Select(x => new TodoResponseDto
         {
@@ -52,6 +62,11 @@ public class TodoController : ControllerBase
 
         if (todo == null)
             return NotFound();
+
+        if (CurrentUserRole != "Admin" && todo.UserId != CurrentUserId)
+        {
+            return Forbid();
+        }
 
         var response = new TodoResponseDto
         {
@@ -83,7 +98,7 @@ public class TodoController : ControllerBase
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> Create([FromBody] TodoCreateDto dto)
     {
-        var item = await _service.CreateAsync(dto.Title);
+        var item = await _service.CreateAsync(dto.Title, CurrentUserId);
 
         var response = new TodoResponseDto
         {
@@ -105,6 +120,15 @@ public class TodoController : ControllerBase
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> Update(int id, [FromBody] TodoResponseDto dto)
     {
+        var todo = await _service.GetByIdAsync(id);
+
+        if (todo == null)
+            return NotFound();
+        if (CurrentUserRole != "Admin" && todo.UserId != CurrentUserId)
+        {
+            return Forbid();
+        }
+
         var result = await _service.UpdateAsync(id, dto.Title, dto.IsCompleted);
 
         if (!result)
@@ -123,6 +147,16 @@ public class TodoController : ControllerBase
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> Delete(int id)
     {
+        var todo = await _service.GetByIdAsync(id);
+
+        if (todo == null)
+            return NotFound();
+
+        if (CurrentUserRole != "Admin" && todo.UserId != CurrentUserId)
+        {
+            return Forbid();
+        }
+
         var result = await _service.DeleteAsync(id);
 
         if (!result)
@@ -130,4 +164,6 @@ public class TodoController : ControllerBase
 
         return NoContent();
     }
+    private int CurrentUserId => int.Parse(User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value ?? "0");
+    private string CurrentUserRole => User.FindFirst(System.Security.Claims.ClaimTypes.Role)?.Value ?? string.Empty;
 }
